Moved files back into correct directory for correct casing
